Adding in the 'test' xtm files
Finally managed to get a pleasant sound out of the thing. Right now I feel like the best approach is going to be to build modules so I can work at a higher level and think more in terms of music. 

I think there's actually some stuff already in the framework that isn't documented which would already help with working at a higher level – e.g. the pitch class module, the metronome, etc. I really want to learn this API kind of stuff even if i have to reverse engineer and document it myself.